Reformatted PNPSolverNode
